Revert "Method level suppressions (#2)"
This reverts commit 1f33882dab9805e83e7464e7f8e9b7d59bac3aca.
