+++ modernizer-maven-plugin/src/test/java/org/gaul/modernizer_maven_plugin/ModernizerTest.java
@@ -88,12 +88,16 @@ public final class ModernizerTest {
             Collections.<Pattern>emptySet();
     private static final Collection<String> NO_IGNORED_PACKAGES =
             Collections.<String>emptySet();
-    private static final Collection<String> NO_IGNORED_METHODS =
-            Collections.<String>emptySet();
 
     @Before
     public void setUp() throws Exception {
-        violations = ModernizerTestUtils.readViolations();
+        InputStream is = Modernizer.class.getResourceAsStream(
+                "/modernizer.xml");
+        try {
+            violations = Modernizer.parseFromXml(is);
+        } finally {
+            Utils.closeQuietly(is);
+        }
     }
 
     @Test
@@ -201,8 +205,7 @@ public final class ModernizerTest {
                 "java/lang/String.getBytes:(Ljava/lang/String;)[B");
         Collection<ViolationOccurrence> occurrences = new Modernizer(
                 "1.6", violations, exclusions, NO_EXCLUSION_PATTERNS,
-                NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS,
-                NO_IGNORED_METHODS).check(cr);
+                NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS).check(cr);
         assertThat(occurrences).hasSize(0);
     }
 
@@ -214,8 +217,7 @@ public final class ModernizerTest {
                 Pattern.compile("java/lang/.*"));
         Collection<ViolationOccurrence> occurrences = new Modernizer(
                 "1.6", violations, NO_EXCLUSIONS, exclusionPatterns,
-                NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS,
-                NO_IGNORED_METHODS).check(cr);
+                NO_IGNORED_PACKAGES, NO_EXCLUSION_PATTERNS).check(cr);
         assertThat(occurrences).hasSize(0);
     }
 
@@ -227,8 +229,7 @@ public final class ModernizerTest {
                 StringGetBytesString.class.getPackage().getName());
         Collection<ViolationOccurrence> occurrences = new Modernizer(
                 "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-                ignorePackages, NO_EXCLUSION_PATTERNS,
-                NO_IGNORED_METHODS).check(cr);
+                ignorePackages, NO_EXCLUSION_PATTERNS).check(cr);
         assertThat(occurrences).hasSize(0);
     }
 
@@ -239,8 +240,7 @@ public final class ModernizerTest {
         Collection<String> ignorePackages = Collections.singleton("org.gaul");
         Collection<ViolationOccurrence> occurrences = new Modernizer(
                 "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-                ignorePackages, NO_EXCLUSION_PATTERNS,
-                NO_IGNORED_METHODS).check(cr);
+                ignorePackages, NO_EXCLUSION_PATTERNS).check(cr);
         assertThat(occurrences).hasSize(0);
     }
 
@@ -251,8 +251,7 @@ public final class ModernizerTest {
         Collection<String> ignorePackages = Collections.singleton("org");
         Collection<ViolationOccurrence> occurrences = new Modernizer(
                 "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-                ignorePackages, NO_EXCLUSION_PATTERNS,
-                NO_IGNORED_METHODS).check(cr);
+                ignorePackages, NO_EXCLUSION_PATTERNS).check(cr);
         assertThat(occurrences).hasSize(0);
     }
 
@@ -263,8 +262,7 @@ public final class ModernizerTest {
         Collection<String> ignorePackages = Collections.singleton("org.gau");
         Collection<ViolationOccurrence> occurrences = new Modernizer(
                 "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-                ignorePackages, NO_EXCLUSION_PATTERNS,
-                NO_IGNORED_METHODS).check(cr);
+                ignorePackages, NO_EXCLUSION_PATTERNS).check(cr);
         assertThat(occurrences).hasSize(1);
     }
 
@@ -276,8 +274,7 @@ public final class ModernizerTest {
                 Pattern.compile(".*StringGetBytesString"));
         Collection<ViolationOccurrence> occurrences = new Modernizer(
                 "1.6", violations, NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS,
-                NO_IGNORED_PACKAGES, ignoreClassNamePatterns,
-                NO_IGNORED_METHODS).check(cr);
+                NO_IGNORED_PACKAGES, ignoreClassNamePatterns).check(cr);
         assertThat(occurrences).hasSize(0);
     }
 
@@ -309,7 +306,7 @@ public final class ModernizerTest {
                 new Violation(name, 5, ""));
         Modernizer modernizer = new Modernizer("1.5", testViolations,
                 NO_EXCLUSIONS, NO_EXCLUSION_PATTERNS, NO_IGNORED_PACKAGES,
-                NO_EXCLUSION_PATTERNS, NO_IGNORED_METHODS);
+                NO_EXCLUSION_PATTERNS);
         ClassReader cr = new ClassReader(AnnotatedMethod.class.getName());
         Collection<ViolationOccurrence> occurences =
                 modernizer.check(cr);
@@ -379,7 +376,7 @@ public final class ModernizerTest {
     private Modernizer createModernizer(String javaVersion) {
         return new Modernizer(javaVersion, violations, NO_EXCLUSIONS,
                 NO_EXCLUSION_PATTERNS, NO_IGNORED_PACKAGES,
-                NO_EXCLUSION_PATTERNS, NO_IGNORED_METHODS);
+                NO_EXCLUSION_PATTERNS);
     }
 
     private static class CharsetsTestClass {
